feat(python): allow inline expressions in asof_join
